Added summary email for backups that have occurred
The summary consists of all of the backups that have occurred since the
last scheduled backup summary email.  This can be configured and is a
schedule kind of similar to the schedule for backing up machines.
Also included in this commit is the db structure for the backupEvent as
this was needed for the summary email.