Renewed dependancies
Most updated packages (even major versions) appear to be compatible.
The exception to this is log4js. We need to heavily modify our log4js
setup so that we can support the latest version (2.x). This should be a
separate item, however, as it is a larger change than I am prepared to
push at this time.