Sweeping change to handling config files.
Added a validation to the config reading, but in researching
implementation for this a lot of tangentially related items came up
including the way we handle defaults, the NODE_ENV variable, consistency
of the items in the example config and the application, what paths and
values can be assumed (as this is intended as a system service), etc.

A lot of this was handled, but more cleanup will follow.

Immediate next steps:
1) Adjust logging initialization to support new log4js (>=2.0)
2) Ensure all config values are used appropriately