Fixed next schedule generation
I had to add handling for milliseconds, as the new dates being generated
were using the same milliseconds as the current timestamp. Therefore,
the comparison for whether or not a date had already passed was wrong
for the first 999 milliseconds after the timeout. Typically we would see
a few calls occur within this time period.