Reorganized config transformations.
I moved the transform of the rsync include/exclude settings into the
prepareMachines() function as it really is a better location for that
logic. In addition, the machine configs in the config object no longer
get transformed (only on the `machine` Object) to better reflect the
difference between the config as the user wrote it and our internal
changes that are needed for rsync to operate the way we want.