Removed specific errmsg content & 'notifications'
This function now takes in a config object with just the notifications
attributes, thus there is no need to create a notifications attribute
for the object being passed inside.

Also removed the specific error message from the tests, as these could
change and it would be hard to update the message in each test for each
edit.