Upgraded Jest.
Reconfigured to specify a single test root. This turned out to be
required for my workstation or else the test suite would simply hang. I
am not sure if this is related to the upgrade or not, but since all our
test scripts live in a single directory anyway, I believe this is a safe
configuration unless someone objects.