Updated the machine specific page
This is not the final version of this page, but enough to enable me to
create an alpha that I can use for our install script. This at least
gives the full and complete history of the backups from the event
database, but doesn't show which are retained on disk and which have
been deleted. Also the schedule isn't represented at all. Lots of
room for improvement in this area.