Better HTTPS support to encourage sane configs.
I've decided to change the goal for #35 as I think it most situations,
the admin will either only want to setup the HTTPS server, or else use a
forward proxy to achieve HTTPS security. So we will not open up both
ports in either case. I did extensively test this setup and it seems to
work as I want, complete with warnings to the user for unexpected setups
that might lead to broken session handling.