Migrated some intense disk operations to Promises.
This change comes after a lot of background research into using native
Node FS operations to count disk space instead of relying on external
tools (df / du) from GNU coreutils. It turns out the external calls are
the best for the performance of this project, and as this can be an
expensive operation we don't want to take the hit just to reduce the
dependencies on this package. This is only partially completed as a few
places that relied on the old code are not ready for Promises yet. In
addition, much of the affected code will be changed once we complete
support for the recently updated memory models.