Fixed write permission on 'npm install'
Also switched the config to a bind-mount for two way data propagation.

Although the docker-compose environment is fully functional with this
fix, I am still contemplating some changes. Namely because of the empty
root owned directories that get created on the host (etc/node_modules).
In the future we might adjust the project layout for better development,
if we can think of improvements. For now this will be sufficient.