Fixed data sync between disk and memory
I really don't like the memory model implementation with regards to the
buckets at this point in time. Rather than making a major change, I
fixed the code so it is working, but in the future we will need to look
closely at the design of this aspect of the system and improve things.

All of the mutable memory models are getting confusing, especially with
the unclear definition of responsibilities for all of the various
methods in the system. A full reworking of these should be possible.